// src/enums.ts
export enum IperfRunnerState {
  Unknown,
  ready,
  initialising,
  running,
  error,
  stopping,
  finished,
}

export enum IperfState {
  TEST_START = 1,
  TEST_RUNNING = 2,
  TEST_END = 4,
  PARAM_EXCHANGE = 9,
  CREATE_STREAMS = 10,
  SERVER_TERMINATE = 11,
  CLIENT_TERMINATE = 12,
  EXCHANGE_RESULTS = 13,
  DISPLAY_RESULTS = 14,
  IPERF_START = 15,
  IPERF_DONE = 16,
  ACCESS_DENIED = -1,
  SERVER_ERROR = -2,
  UNKNOWN = 0,
}

export enum IperfProtocol {
  TCP = 'Ptcp',
  UDP = 'Pudp',
  SCTP = 'Psctp',
}

export enum IperfRole {
  SERVER = 115,
  CLIENT = 99,
}

export enum IperfDirection {
  DOWNLOAD = 1,
  UPLOAD = 0,
}

export enum IperfError {
  UNKNOWN = -1,
  IENONE = 0,
  /* Parameter errors */
  IESERVCLIENT = 1,
  IENOROLE = 2,
  IESERVERONLY = 3,
  IECLIENTONLY = 4,
  IEDURATION = 5,
  IENUMSTREAMS = 6,
  IEBLOCKSIZE = 7,
  IEBUFSIZE = 8,
  IEINTERVAL = 9,
  IEMSS = 10,
  IENOSENDFILE = 11,
  IEOMIT = 12,
  IEUNIMP = 13,
  IEFILE = 14,
  IEBURST = 15,
  IEENDCONDITIONS = 16,
  IELOGFILE = 17,
  IENOSCTP = 18,
  IEBIND = 19,
  IEUDPBLOCKSIZE = 20,
  IEBADTOS = 21,
  IESETCLIENTAUTH = 22,
  IESETSERVERAUTH = 23,
  IEBADFORMAT = 24,
  IEREVERSEBIDIR = 25,
  IEBADPORT = 26,
  IETOTALRATE = 27,
  IETOTALINTERVAL = 28,
  /* Test errors */
  IENEWTEST = 100,
  IEINITTEST = 101,
  IELISTEN = 102,
  IECONNECT = 103,
  IEACCEPT = 104,
  IESENDCOOKIE = 105,
  IERECVCOOKIE = 106,
  IECTRLWRITE = 107,
  IECTRLREAD = 108,
  IECTRLCLOSE = 109,
  IEMESSAGE = 110,
  IESENDMESSAGE = 111,
  IERECVMESSAGE = 112,
  IESENDPARAMS = 113,
  IERECVPARAMS = 114,
  IEPACKAGERESULTS = 115,
  IESENDRESULTS = 116,
  IERECVRESULTS = 117,
  IESELECT = 118,
  IECLIENTTERM = 119,
  IESERVERTERM = 120,
  IEACCESSDENIED = 121,
  IESETNODELAY = 122,
  IESETMSS = 123,
  IESETBUF = 124,
  IESETTOS = 125,
  IESETCOS = 126,
  IESETFLOW = 127,
  IEREUSEADDR = 128,
  IENONBLOCKING = 129,
  IESETWINDOWSIZE = 130,
  IEPROTOCOL = 131,
  IEAFFINITY = 132,
  IEDAEMON = 133,
  IESETCONGESTION = 134,
  IEPIDFILE = 135,
  IEV6ONLY = 136,
  IESETSCTPDISABLEFRAG = 137,
  IESETSCTPNSTREAM = 138,
  IESETSCTPBINDX = 139,
  IESETPACING = 140,
  IESETBUF2 = 141,
  IEAUTHTEST = 142,
  /* Stream errors */
  IECREATESTREAM = 200,
  IEINITSTREAM = 201,
  IESTREAMLISTEN = 202,
  IESTREAMCONNECT = 203,
  IESTREAMACCEPT = 204,
  IESTREAMWRITE = 205,
  IESTREAMREAD = 206,
  IESTREAMCLOSE = 207,
  IESTREAMID = 208,
  /* Timer errors */
  IENEWTIMER = 300,
  IEUPDATETIMER = 301,
  INIT_ERROR = 400,
  INIT_ERROR_DEFAULTS = 401,
}